+++ packages/driver-typescript/src/types.ts
@@ -1,4 +1,5 @@
-import typescript from 'typescript';
+import ts from 'typescript';
+import { DriverOptions } from '@beemo/core';
 
 export type TargetSetting = 'ES5' | 'ES2015' | 'ES2016' | 'ES2017' | 'ES2018' | 'ESNEXT';
 
@@ -46,12 +47,13 @@ export type JSXSetting = 'preserve' | 'react-native' | 'react';
 
 export interface TypeScriptConfig {
   compileOnSave?: boolean;
-  compilerOptions?: typescript.CompilerOptions;
+  compilerOptions?: ts.CompilerOptions;
   exclude?: string[];
   extends?: string;
   files?: string[];
   include?: string[];
-  typeAcquisition?: typescript.TypeAcquisition;
+  references?: ts.ProjectReference[];
+  typeAcquisition?: ts.TypeAcquisition;
 }
 
 export interface TypeScriptArgs {
@@ -97,4 +99,12 @@ export interface TypeScriptArgs {
   version?: boolean;
   w?: boolean;
   watch?: boolean;
+}
+
+export interface TypeScriptOptions extends DriverOptions {
+  buildFolder: string;
+  globalTypes: boolean;
+  srcFolder: string;
+  testsFolder: string;
+  typesFolder: string;
 }