+++ packages/driver-typescript/CHANGELOG.md
@@ -3,3 +3,4 @@
 #### 🎉 Release
 
 - Initial release!
+- Added automatic project references support via the `--reference-workspaces` option.